API/requirements.txt
requests
keras
numpy
tensorflow
fastapi
Pillow
gunicorn
uvicorn
protobuf
